+++ DependencyInjection/HWIOAuthExtension.php
@@ -16,6 +16,7 @@ use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\DefinitionDecorator;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
+use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
 /**
@@ -65,7 +66,7 @@ class HWIOAuthExtension extends Extension
         if (isset($config['fosub'])) {
             $container
                 ->setDefinition('hwi_oauth.user.provider.fosub_bridge', new DefinitionDecorator('hwi_oauth.user.provider.fosub_bridge.def'))
-                ->addArgument($container)
+                ->addArgument(new Reference('security.context'))
                 ->addArgument($config['fosub']['properties'])
             ;
         }